pyproject.toml
[tool.poetry]
name = "mangawatch"
version = "0.1.0"
description = ""
authors = ["arjoonn sharma <arjoonn.94@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
python-telegram-bot = "^12.8"
requests = "^2.24.0"
bs4 = "^0.0.1"
lxml = "^4.5.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
